Add support for automatically registering a consumer
In most use cases, Henson-AMQP is being used for its ``Consumer`` and
``Producer`` classes. Currently, Henson applications using this
``Consumer`` class require boilerplate to create and assign a consumer.
This change reduces that boilerplate to a setting.

Initially, this feature is disabled by default to prevent existing
applications from inadvertently creating multiple consumers after
upgrading to a new release of Henson-AMQP. In the future, if widely
used, it may become enabled by default.